requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-Login>=0.6
python-dotenv>=1.0
psycopg2-binary>=2.9
pymysql
cryptography